Misc: Util: Use user's real $HOME with fallback to per-user temp (bnc#780973)
When making changes as a regular (non-hacluster) user, in some cases
the ownership/permissions of $HOME/.cache and $HOME/.config cause the
crm shell invocation to fail.  This fix uses the actual user's home
directory, or a specific per-user fallback temp dir if the case there's
no actual home directory.

Signed-off-by: Tim Serong <tserong@suse.com>